requirements.txt
# Core Python requirements
requests
aiohttp
beautifulsoup4
lxml
html5lib

# Reddit API client
praw

# LLM Analysis 
openai  # Only if you use OpenAI-compatible APIs
# If you use Azure OpenAI, you can also use azure-core, azure-identity if needed

# NLP and processing
nltk
langdetect
scikit-learn
numpy

# Data handling
python-dotenv
tqdm

# JSON schema and validation (for schema manager, optional)
jsonschema

# Logging, utils
loguru

# For testing and mocking (optional but useful)
pytest
pytest-mock

# Only needed if you do transformations like Base64, encoding, etc.
# base64 is built-in, so no need to list it separately
praw>=7.7.0
openai>=1.3.0
anthropic>=0.5.0
python-dotenv>=1.0.0
tabulate>=0.9.0
azure-ai-inference>=1.0.0b9


# Core dependencies
tweepy>=4.14.0
networkx>=3.2.1
python-dotenv>=1.0.0
pandas>=2.0.3
pydantic>=1.10.7
numpy>=1.24.3
scipy>=1.10.1

# LLM support (Azure OpenAI)
openai>=1.14.2
azure-ai-inference>=1.0.0b9
azure-core>=1.29.5

# Optional: For CLI output formatting
tabulate>=0.9.0

discord.py==2.3.2
aiohttp==3.9.3
audioop-lts==0.2.1
PyNaCl==1.5.0

#coordinator

sentence-transformers==2.2.2
networkx==2.8.4
pandas==1.5.0
scikit-learn==1.1.2
torch==1.13.0

#Twitter Agent
# Core dependencies
tweepy>=4.14.0
networkx>=3.2.1
python-dotenv>=1.0.0
pandas>=2.0.3
pydantic>=1.10.7
numpy>=1.24.3
scipy>=1.10.1

# LLM support (Azure OpenAI)
openai>=1.14.2
azure-ai-inference>=1.0.0b9
azure-core>=1.29.5

# Optional: For CLI output formatting
tabulate>=0.9.0


#Reddit Agent
praw
openai
anthropic
python-dotenv
tabulate
azure-ai-inference


#GitHub Agent 
# Core dependencies
PyGithub==1.59.0
aiohttp==3.9.5
networkx==3.3
requests==2.31.0

# Optional LLM integration tools
openai==1.24.1  # Only if using OpenAI SDK for non-Azure endpoints
# langchain==0.1.16  # Uncomment if you plan to use LangChain logic

# Code parsing (future use)
tree-sitter==0.20.1

# Utility
python-dotenv==1.0.1  # Optional: for managing secrets via .env

# Minimum Python version: 3.8+


#Twitter Agent
aiohttp
beautifulsoup4
langdetect
iso-639
langchain
tqdm


#Discord Agent
discord.py==2.3.2
aiohttp==3.9.3
openai>=1.13.3
nltk==3.8.1
numpy==1.26.4
audioop-lts==0.2.1
PyNaCl==1.5.0



